requirements/requirements.txt
# Librerias
python==3.12.3
pyserial==3.5